+++ prefs.js
@@ -12,10 +12,6 @@
 
 'use strict';
 
-import GIRepository from 'gi://GIRepository';
-GIRepository.Repository.prepend_search_path('/usr/lib64/gnome-shell');
-GIRepository.Repository.prepend_library_path('/usr/lib64/gnome-shell');
-
 import Adw from 'gi://Adw';
 import GObject from 'gi://GObject';
 import Gio from 'gi://Gio';
@@ -1180,5 +1176,7 @@ export default class EasyScreenCastPreferences extends ExtensionPreferences {
         group.add(widget);
 
         window.add(page);
+        // fix for #360 - make it a bit wider to show the "X" to close the window
+        window.set_default_size(window['default-width'] + 50, window['default-height']);
     }
 }